conversion precision warning
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@11697 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/common/GameTime.cxx
@@ -197,7 +197,7 @@ void GameTime::setStepTime()
     stepTime = thisStep;
   } else {
     // long term prediction
-    const double diffLocal = thisStep - avgPoint.localTime;
+    const double diffLocal = (double)(thisStep - avgPoint.localTime);
     const double longPred = (double)avgPoint.netTime + (diffLocal * avgRate);
     // short term prediction
     const double skipTime = (double)(thisStep - lastStep);
